finish setup new tm expansion

+++ include/pokemon.h
@@ -577,8 +577,8 @@ u8 CheckPartyPokerus(struct Pokemon *party, u8 selection);
 u8 CheckPartyHasHadPokerus(struct Pokemon *party, u8 selection);
 void PartySpreadPokerus(struct Pokemon *party);
 bool8 TryIncrementMonLevel(struct Pokemon *mon);
-u32 CanMonLearnTMHM(struct Pokemon *mon, u8 tm); //change to u16 for both, later
-u32 CanSpeciesLearnTMHM(u16 species, u8 tm);
+u32 CanMonLearnTMHM(struct Pokemon *mon, u16 tm); //change to u16 for both, later
+u32 CanSpeciesLearnTMHM(u16 species, u16 tm);
 u8 GetMoveRelearnerMoves(struct Pokemon *mon, u16 *moves);
 u8 GetLevelUpMovesBySpecies(u16 species, u16 *moves); //keep track of this for rotom forms 
 u8 GetNumberOfRelearnableMoves(struct Pokemon *mon);//this too
